fixed compile errors and warnings using Dev-C++ in "Compile as C++" mode
git-svn-id: svn://svn.icculus.org/twilight/trunk/darkplaces@9198 d7cf8633-e32d-0410-b094-e92efae38249

+++ menu.c
@@ -5377,7 +5377,7 @@ void MR_Init(void)
 			}
 			if(video_resolutions[i].conwidth > video_resolutions[i].width || video_resolutions[i].conheight > video_resolutions[i].height)
 			{
-				double f1, f2;
+				int f1, f2;
 				f1 = video_resolutions[i].conwidth > video_resolutions[i].width;
 				f2 = video_resolutions[i].conheight > video_resolutions[i].height;
 				if(f1 > f2)
